Style(main): Detach header file, make style consistent
